requirements.txt
streamlit==1.29.0
supabase==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
python-dotenv==1.0.0
httpx==0.24.1
colorlog==6.7.0
